+++ src/core/managers/__tests__/ModelManager.test.ts
@@ -4,7 +4,7 @@ import { IFloorAlignmentService } from '../../services/IFloorAlignmentService';
 import { ISceneSetupService } from '../../services/ISceneSetupService';
 import { ThreeViewerError, ErrorCode } from '../../../errors';
 import { TypedEventEmitter } from '../../../events/EventEmitter';
-import { ViewerEventMap } from '../../events/ViewerEvents';
+import { ViewerEventMap } from '../../events/CoreViewerEvents';
 
 describe('ModelManager', () => {
   let modelManager: ModelManager;